+++ ncplib/__init__.py
@@ -52,8 +52,19 @@ More information
 """
 
 
-from ncplib.client import *  # noqa
-from ncplib.connection import *  # noqa
-from ncplib.errors import *  # noqa
-from ncplib.server import *  # noqa
-from ncplib.values import *  # noqa
+from ncplib.client import connect as connect  # noqa
+from ncplib.connection import Connection as Connection, Response as Response, Field as Field  # noqa
+from ncplib.errors import (  # noqa
+    NCPError as NCPError,
+    NetworkError as NetworkError,
+    AuthenticationError as AuthenticationError,
+    NetworkTimeoutError as NetworkTimeoutError,
+    ConnectionClosed as ConnectionClosed,
+    CommandError as CommandError,
+    DecodeError as DecodeError,
+    NCPWarning as NCPWarning,
+    CommandWarning as CommandWarning,
+    DecodeWarning as DecodeWarning,
+)
+from ncplib.server import start_server as start_server  # noqa
+from ncplib.values import u32 as u32, i64 as i64, u64 as u64, f64 as f64  # noqa